Initial import of p5-AnyEvent-HTTP version 1.44 in the NetBSD Packages Collection.
The Perl 5 module AnyEvent::HTTP is an AnyEvent user, you need to
make sure that you use and run a supported event loop.  It implements
a simple, stateless and non-blocking HTTP client. It supports GET,
POST and other request methods, cookies and more, all on a very low
level. It can follow redirects supports proxies and automatically
limits the number of connections to the values specified in the
RFC.