0.22: Avoid infinite loop in some error cases.
Sometimes, the set of packages to be replaced ($REPLACE_TODO) does not
have any members which are installed (and hence in $TSORTED).  This
caused the loop over $TSORTED to not find any packages to replace.
The check for not having found one was missing, resulting in the last
package in TSORTED being replaced.  Because $REPLACE_TODO is still
non-empty, this repeats.

This patch adds an abort when no packages in $TSORTED are in
$REPLACE_TODO.  Analysis and patch from Ian D. Leroux on
pkgsrc-users@.