Update Fossil to release 2011-05-23 15:11:12
New in release 2011-05-23 15:11:12:

This release merges in the windows internationalization patches.
Fossil should now work better on windows machines that use a
non-ASCII and non-UTF8 code page for the DOS box.


New in release 2011-05-12 14:56:52:

This release adds an enhanced configuration sync capability
which entails an irreversible schema change.
You _must_ run "fossil rebuild" on all of your repositories after
updating to this version of fossil.

Other changes in this release include:

  * Refactor the "add", "rm", and "addremove" commands to
    simplify the code and fix various problems.
  * Added a "diff" hyperlink after each file in the "Show Files"
    timeline view.
  * The "fossil open" and "fossil co" commands always prompt
    before overwriting preexisting files unless the --force
    option is used.
  * Enhanced the merge-conflict markup to show both recent
    versions and the common-ancestor version.
  * Change the definition of what it means to be a "leaf"
    check-in, to be consistent and to work better for most
    people.
  * Commands that recursively decend through the file hierarchy
    ("fossil extra", "fossil clean", etc.) will now ignore
    nested checkouts.
  * Automatically delete the _FOSSIL_ file upon a failed "open".
  * Improvements to the "annotate" feature.
  * Other minor bug fixes.