audio/xmms-cdread: Support DragonFly
There were a few issues with this package wrt DragonFly. The biggest issue
is the use of ETIME, which should not be used with sockets.  It doesn't
even appear to be used later.  NetBSD defines ETIME, but neither DragonFly
nor FreeBSD even have this definition in errno.  Somebody might want to
look at this choice for the variable.

Libtool also choked due to a missing tag, the FreeBSD volume control was
needed, and some standard headers were added.