finances/kmymoney2: Fix all platforms
The main problem with this package is that sqlite2 was building
automatically, but the makefile for it was broken.  The QTDIR variable
wasn't getting expanded in the makefile, so the rpath for sqlite3 library
was corrupt.  Once the Makefile was fixed, libsqlite3.so built normally,
but it didn't get installed due to another Makefile issue.  Finally, it
built and installed, but then the PLIST was missing its entry.

After everything was verified building ok, I discovered somebody had
added and then commented out "sqlite3" option.  Since it's building by
default, I took this option out.  If somebody else decides that it
should be optional, I'll let them tweak the configure args, PLIST, and
options file to make that happen.