1.4.8: Feature and bugfix release. New Touch commands, counters, and a change to connection limit functionality. *Touch Commands *Binary Touch/GAT commands were backported from 1.6. *New GATK/GATKQ commands were added for completeness. *Finally, an Ascii protocol touch command was also added. *Fast Connection Limit Handling *Internal Hash Table *New Stats: expired_unfetched, evicted_unfetched
1.4.9:
Small bugfix release. Mainly fixing a critical issue where using -c to increase the connection limit was broken in 1.4.8. If you are on 1.4.8, an upgrade is highly recommended.

1.4.10:
This release is focused on thread scalability and performance improvements. This release should be able to feed data back faster than any network card can support as of this writing.

1.4.11:
Memcached 1.4.11. Fixes race conditions and crashes introduced in 1.4.10. Adds the ability to rebalance and reassign slab memory.
*Slab Reassign
*Slab Automove
*New Stats: slab_reassign_running, slabs_moved

1.4.12:
Fix a small number of bugs, mostly in building on different platforms.

1.4.13:
Really tiny release with some important build fixes which were accidentally omitted from 1.4.12.