emulators/handy_sdl: Really fix normal user building
The post-extract target on this package was supposed to allow normal users
to build this package, but it didn't work for me.  It may have been the
attempt at changing $WRKSRC instead of $WRKSRC/src (recursively).
The updated command touches only directories rather than all files and
now handy_sdl builds for me with a non-root user.