Update to 1.2.4.
1.2.4:
Alan Coopersmith (3):
      Make sure to leave room for trailing nil byte in yyGetNumber
      Delete redundant redeclarations of extern Atoms in symbols.c
      xkbcomp 1.2.4

Andreas Wettstein (1):
      include resets group compatibility modifiers #43091

Jeremy Huddleston (6):
      Include strings.h for strcasecmp and string.h for strdup
      Add missing _X_NORETURN to uFatalError
      Fix -Wformat warnings
      Correct our zeroing out of the row in InitRowInfo
      Silence -Wshadow warnings
      Silence -Wparentheses-equality

Peter Hutterer (2):
      Print version number on -version.
      Add .pc file for version export

1.2.3:
There was a pretty bad problem with the 1.2.2 release, which quite
effectively crushed Any+AnyOfOrNone(All) interp mappings, in that
xkeyboard-config actually relied on an explicit Any+AnyOfOrNone(All)
mapping to function.  1.2.3 builds on the previous 1.2.2 fix to allow
explicit Any+AnyOfOrNone(All) mappings, while still ignoring interp
mappings whose keysyms could not be found.

I've regression-tested this against xkbcomp 1.1.1 by comparing the
results of compiling every layout (primary variants only) in
xkeyboard-config, and there were no differences at all.

Distributions are strongly recommended to push 1.2.3 instead of 1.2.2,
whose tarballs have been removed from the archive.

1.2.2:
This is an semi-urgent bugfix release for the two commits of mine, which
fix a particularly bad failure in the compat code.  If you had a
SymInterpret definition (i.e. to map a keysym to an action, as used for
VT switch and others) with an unknown keycode, every key that wasn't
already mapped to an action would get mapped to your new action.

So, for example, if you had XF86LogWinTree mapped to the PrWins action,
and a libX11 that was unaware of the new symbol, every key aside from VT
switching, zapping, pointer keys and modifiers, would dump the window
tree to your X log, and do nothing else.

Distributions are strongly, strongly recommended to upgrade, including
in stable series.

1.2.1:
This release allows the uploading of keymaps per-keyboard using the -i flag
(which wasn't hooked up to anything before).

Example usage:
   xkbcomp -i 12 qwerty.xkb :0
   xkbcomp -i 13 azerty.xkb :0

The above will load a qwerty and azerty map in to the keyboards 12 and 13,
respectively. Watch out for bug 21859 though.
http://bugs.freedesktop.org/show_bug.cgi?id=21859

Aside from this new feature, we have the usual cleanups in the build system.

1.2.0:
This release provides performance improvements and bug fixes, mixed in
with the usual recent collection of build configuration improvements and
janitorial cleanups.